refactor(library): avoid stuttering, balance hier
1. avoid name stuttering in modules withing the iqb package

2. balance test hier by using ./test/iqb
